+++ Changelog.md
@@ -13,6 +13,7 @@
     a) shipping one order in one shipment  
     b) shipping one order in multiple shipments  
     c) shipping multiple orders in one shipment  
+- Added `Shipment::orders` and `Order::shipments` to the Foundation models
 - Changed the random number algo in `TimeHash` generator from `mt_rand` to `random_int` for a decreased collision probability
 - Fixed the `Carrier::name()` method to overcome accessor infinite loop
 - Fixed the address creation with order factory when passing an explicit address type